requirements.txt
# Web框架
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Werkzeug==2.3.7

# 数据库
PyMySQL==1.0.0
cryptography==3.4.0
greenlet==3.0.1

# 云服务
cos-python-sdk-v5==1.9.38
volcengine-python-sdk[ark]==1.0.0

# 工具包
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.0.0

# 生产环境服务器
gunicorn==21.2.0